requirements.txt
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pymongo>=4.0.0
